+++ docs/src/data/links.tsx
@@ -41,6 +41,7 @@ import {
   MdChevronRight,
   MdPanToolAlt,
   MdLabel,
+  MdAccountCircle,
 } from 'react-icons/md';
 
 export interface ComponentNavItem {
@@ -211,9 +212,9 @@ export const connectedComponents: ComponentNavItem[] = [
     tertiary: true,
   },
   {
-    href: '/connected-components/storage/storagemanager',
-    label: 'Storage Manager',
-    body: 'StorageManager component allows users to upload and manage files in your Amplify backend.',
+    href: '/connected-components/storage/fileuploader',
+    label: 'File Uploader',
+    body: 'FileUploader component allows users to upload and manage files in your Amplify backend.',
     platforms: ['react'],
     tertiary: true,
   },
@@ -226,6 +227,13 @@ export const connectedComponents: ComponentNavItem[] = [
 ];
 
 export const dataDisplayComponents: ComponentNavItem[] = [
+  {
+    href: '/components/avatar',
+    label: 'Avatar',
+    body: `An Avatar represents a user by displaying a profile image, initials, or a fallback icon.`,
+    platforms: ['react'],
+    icon: MdAccountCircle,
+  },
   {
     href: '/components/badge',
     label: 'Badge',