+++ packages/react/src/primitives/TextArea/index.ts
@@ -1,1 +1,2 @@
 export { TextArea } from './TextArea';
+export { AutoresizeTextArea } from './AutoresizeTextarea';