+++ packages/react-geo/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @aws-amplify/ui-react-geo
 
+## 2.0.22
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aws-amplify/ui-react-core@3.0.21
+
+## 2.0.21
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aws-amplify/ui-react-core@3.0.20
+
+## 2.0.20
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @aws-amplify/ui-react-core@3.0.19
+
 ## 2.0.19
 
 ### Patch Changes